requirements.txt
Django==4.2
django-bootstrap5==23.1
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg2==2.9.6
requests==2.31.0
Pillow==10.0.0